requirements.txt
streamlit
pypdf
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
